+++ src/modules/apigw/index.ts
@@ -82,7 +82,12 @@ export default class Apigw {
 
   /** 部署 API 网关 */
   async deploy(inputs: ApigwDeployInputs) {
-    const { environment = 'release' as const, oldState = {}, isInputServiceId = false } = inputs;
+    const {
+      environment = 'release' as const,
+      oldState = {},
+      isInputServiceId = false,
+      isAutoRelease = true,
+    } = inputs;
     if (isInputServiceId) {
       return this.deployWIthInputServiceId(inputs as ApigwDeployWithServiceIdInputs);
     }
@@ -107,7 +112,9 @@ export default class Apigw {
       environment,
     });
 
-    await this.service.release({ serviceId, environment });
+    if (isAutoRelease) {
+      await this.service.release({ serviceId, environment });
+    }
 
     console.log(`Deploy service ${serviceId} success`);
 
@@ -162,6 +169,7 @@ export default class Apigw {
       customDomains,
       usagePlan,
       isRemoveTrigger = false,
+      isAutoRelease = true,
     } = inputs;
 
     // check service exist
@@ -180,12 +188,12 @@ export default class Apigw {
 
     // 定制化需求：如果用户在yaml中配置了 serviceId，则只执行删除 api 逻辑
     // 删除后需要重新发布
-    if (isRemoveTrigger) {
+    if (isRemoveTrigger && isAutoRelease) {
       await this.service.release({ serviceId, environment });
       return;
     }
 
-    // remove usage plan
+    // 删除使用计划
     if (usagePlan) {
       await this.usagePlan.remove({
         serviceId,
@@ -194,7 +202,7 @@ export default class Apigw {
       });
     }
 
-    // 2. unbind all custom domains
+    // 解绑自定义域名
     if (customDomains) {
       for (let i = 0; i < customDomains.length; i++) {
         const curDomain = customDomains[i];
@@ -209,38 +217,21 @@ export default class Apigw {
       }
     }
 
-    if (created === true) {
-      // unrelease service
-      console.log(`Unreleasing service: ${serviceId}, environment ${environment}`);
-      await this.removeRequest({
-        Action: 'UnReleaseService',
+    if (created && isAutoRelease) {
+      await this.service.remove({
         serviceId,
-        environmentName: environment,
+        environment,
       });
-      console.log(`Unrelease service ${serviceId}, environment ${environment} success`);
-
-      // 在删除之前，如果关联了标签，需要先删除标签关联
-      if (detail.Tags && detail.Tags.length > 0) {
-        await this.tagClient.deployResourceTags({
-          tags: [],
-          resourceId: serviceId,
-          serviceType: ApiServiceType.apigw,
-          resourcePrefix: 'service',
-        });
-      }
-
-      // delete service
-      console.log(`Removing service ${serviceId}`);
-      await this.removeRequest({
-        Action: 'DeleteService',
-        serviceId,
-      });
-      console.log(`Remove service ${serviceId} success`);
     }
   }
 
   async deployWIthInputServiceId(inputs: ApigwDeployWithServiceIdInputs) {
-    const { environment = 'release' as const, oldState = {}, serviceId } = inputs;
+    const {
+      environment = 'release' as const,
+      oldState = {},
+      serviceId,
+      isAutoRelease = true,
+    } = inputs;
     inputs.protocols = getProtocolString(inputs.protocols as ('http' | 'https')[]);
 
     const endpoints = inputs.endpoints || [];
@@ -255,7 +246,9 @@ export default class Apigw {
         environment,
       });
 
-      await this.service.release({ serviceId, environment });
+      if (isAutoRelease) {
+        await this.service.release({ serviceId, environment });
+      }
 
       console.log(`Deploy service ${serviceId} success`);